Rename select (#552)
* Single factorization bloqs.

* Use OF test params.

* Better unit tests.

* Better tests.

* Remove unused import.

* Fix QROM data size for SF and add note / reference.

* Remove duplicate num_aux / unused variables.

* Add missing __init__.

* Add missing init / license.

* Format.

* Use select and prepare oracle.

* Format.

* Fix imports.

* Fix argument.

* Fix docstring.

* Sparse SELECT decomposition.

* Temporarily delete sparse notebook.

* Decomposition for sparse prepare.

* Add back notebook.

* Address review comments.

* Fix test.

* Replace select.py with select_bloq.py.

* Errant select.py.

* Remove aspirational factory method.

* Fix notebook errors.

* Fix lint error.

---------

+++ qualtran/bloqs/chemistry/df/double_factorization.py
@@ -46,7 +46,7 @@ from qualtran.bloqs.chemistry.df.prepare import (
     OuterPrepareDoubleFactorization,
     OutputIndexedData,
 )
-from qualtran.bloqs.chemistry.df.select import ProgRotGateArray
+from qualtran.bloqs.chemistry.df.select_bloq import ProgRotGateArray
 from qualtran.bloqs.multi_control_multi_target_pauli import MultiControlPauli
 from qualtran.bloqs.util_bloqs import ArbitraryClifford
 
@@ -324,7 +324,7 @@ class DoubleFactorizationBlockEncoding(Bloq):
 
     @classmethod
     def build_from_coeffs(
-        cls, one_body_ham: NDArray[float], factorized_two_body_ham: NDArray[float]
+        cls, one_body_ham: NDArray[np.float64], factorized_two_body_ham: NDArray[np.float64]
     ) -> 'DoubleFactorizationBlockEncoding':
         """Factory method to build double factorization block encoding given Hamiltonian inputs.
